refactor: template
