+++ apps/example/.env.template
@@ -1,3 +1,3 @@
 DISCORD_TOKEN=
 ENVIRONMENT=PRODUCTION
-DEV_GUILD_IDS=
+DEV_GUILD_ID=
