+++ apps/example/.eslintrc.cjs
@@ -1,0 +1,3 @@
+// @ts-check
+
+module.exports = require("@leomotors/config/eslint").config();
