+++ apps/example/.prettierrc.cjs
@@ -1,0 +1,13 @@
+// @ts-check
+
+module.exports = require("@leomotors/config/prettier").withSortImports({
+  importOrder: [
+    "^dotenv",
+    "^cocoa-discord",
+    "^discord.js",
+    "^@discordjs",
+    "^[a-zA-Z]",
+    "^[.][.]",
+    "^[.]",
+  ],
+});
