+++ apps/example/data/activities.json
