--- templates/hybrid/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@cocoa-discord/template-hybrid",
-  "version": "1.0.0",
-  "description": "Template Discord Bot using Cocoa Discord Utils (Has both message and slash commands)",
-  "main": "dist/bot.js",
-  "type": "module",
-  "scripts": {
-    "build:app": "tsc",
-    "clean": "rm -rf dist/*",
-    "start": "node --es-module-specifier-resolution=node dist/bot.js",
-    "format": "prettier --write \"src/**/*.ts\"",
-    "lint": "eslint src && prettier --check \"src/**/*.ts\""
-  },
-  "devDependencies": {
-    "@trivago/prettier-plugin-sort-imports": "4.2.0",
-    "@typescript-eslint/eslint-plugin": "6.7.0",
-    "@typescript-eslint/parser": "6.7.0",
-    "eslint": "8.49.0",
-    "prettier": "3.0.3",
-    "typescript": "5.2.2"
-  },
-  "dependencies": {
-    "cocoa-discord-utils": "^2.3.0",
-    "discord.js": "^14.13.0",
-    "dotenv": "^16.3.1"
-  }
-}
