--- templates/slash-only/.env.template
@@ -1,3 +1,0 @@
-DISCORD_TOKEN=
-ENVIRONMENT=PRODUCTION
-DEV_GUILD_IDS=
