--- templates/slash-only/.prettierignore
@@ -1,2 +1,0 @@
-*.md
-*.yml
