--- templates/slash-only/README.md
@@ -1,3 +1,0 @@
-# Template
-
-Template Discord Bot for Cocoa Discord Utils
