--- templates/slash-only/data/activities.json
@@ -1,5 +1,0 @@
-{
-    "PLAYING": [
-        "Sample Activity"
-    ]
-}
