--- templates/slash-only/src/commands/styles.ts
@@ -1,9 +1,0 @@
-import { EmbedStyle } from "cocoa-discord-utils";
-
-export * from "./main";
-
-export const style = new EmbedStyle({
-  author: "invoker",
-  // color: 0xFFFF00,
-  // footer: { text: "message you want to put in footer of embed" },
-});
