--- templates/slash-only/src/environment.ts
@@ -1,4 +1,0 @@
-export const GuildIds =
-  process.env.ENVIRONMENT === "PRODUCTION"
-    ? "Global"
-    : process.env.DEV_GUILD_IDS?.split(",") ?? [];
